+++ ipc/requirements.txt
@@ -1,0 +1,3 @@
+theano==0.9
+posix_ipc==1.0
+six
