--- requirements.txt
@@ -1,3 +1,0 @@
-theano==0.9
-trollius
-six
