+++ src/config.h
@@ -38,8 +38,8 @@
 //#define USE_OPENCL
 //#define USE_TUNER
 #define USE_IPC
-// Remember to turn on USE_BLAS and USE_OPENCL when using USE_SERVER_TEST
-// #define USE_IPC_TEST
+// Remember to turn on USE_BLAS, USE_OPENBLAS and USE_OPENCL when using USE_SERVER_TEST
+//#define USE_IPC_TEST
 
 #define PROGRAM_NAME "Leela Zero"
 #define PROGRAM_VERSION "0.6"
